+++ src/types/recursivePartial.ts
@@ -3,9 +3,9 @@
  */
 export type RecursivePartial<T> = 
 {
-    [P in keyof T]?: T[P] extends (infer U)[] ?
-                        RecursivePartial<U>[] : 
-                        T[P] extends object | undefined ?
-                            RecursivePartial<T[P]> :
-                            T[P];
+    [P in keyof T]?: T[P] extends (infer U)[] 
+        ? RecursivePartial<U>[]
+        : T[P] extends object | undefined 
+            ? RecursivePartial<T[P]> 
+            : T[P];
 };